Split OperationOutcome out of errors.js for searchability
