+++ packages/shared-src/src/utils/fhir/index.js
@@ -2,3 +2,4 @@ export * from './datetime';
 export * from './errors';
 export * from './parameters';
 export * from './schemata';
+export * from './OperationOutcome';
